# swagger/swagger40.yaml
swagger: "2.0"
info:
  description: "This is test server for GIRC app"
  version: "3.0.0"
  title: "GIRC App API"
  # termsOfService: "http://swagger.io/terms/"
 # contact:
   # email: "apiteam@swagger.io"
#  license:
  #  name: "Apache 2.0"
  #  url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: girc.app
basePath: "/"
...
schemes:
- "https"
paths:
  /v2/user:
    post:
      tags:
      - "user"
      summary: "register a user"
      description: "register a user with his information"
      operationId: "createUser"
      produces: 
      - "application/json"
      consumes:
      - application/json
      parameters:
      - name: "Authorization"
        in: header
        description: "authorization header contains bearer token"
        type: string
        required: true
      - name: "user"
        in: "body"
        required: true
        description: "a user's information"
        schema:
          $ref: "#/definitions/CreateUser"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/GoodResponse"
        "400":
          description: database error while creating user
          schema:
            $ref: "#/definitions/BadResponse"
        "401":
          description: bad authorization token
          schema:
            $ref: "#/definitions/BadResponse"
    get:
      tags:
      - "user"
      summary: "Get user's information."
      description: "Get a user's information."
      operationId: "getUser"
      produces:
      - "application/json"
      consumes:
      - application/json
      parameters:
      - name: "Authorization"
        in: "header"
        description: "authorization header contains bearer token"
        type: string
        required: true
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/GetUserGoodResponse"
        401:
          description: bad authorization token
          schema: 
            $ref: "#/definitions/BadResponse"
        "404":
          description: "user not found"
          schema:
            $ref: "#/definitions/BadResponse"

    delete:
      tags:
      - user
      summary: unregisters a user, for testing only will be removed in production
      description: deletes a user
      operationId: deleteUser
      produces:
      - "application/json"
      consumes:
      - application/json
      parameters:
      - name: "Authorization"
        in: header
        description: "authorization token"
        type: string
        required: true
      - in: header
        name: User
        description: "contains user id"
        type: string
        required: true
      responses:
        200: 
          description: "successful operation"
          schema:
            $ref: '#/definitions/GoodResponse'
        401:
          description: bad authorization token
          schema:
            $ref: '#/definitions/BadResponse'
        404:
          description: 'user with requested userid not found'
          schema:
            $ref: '#/definitions/BadResponse'
  /v2/incident:
    post:
      tags:
      - "incident"
      summary: "Create an incident"
      description: "use this to create an incident for a userId"
      operationId: "createIncident"
      produces:
      - "application/json"
      consumes:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "authorization header contains bearer token"
        type: string
        required: true
      - in: body
        name: incident
        required: true
        schema:
          $ref: "#/definitions/CreateIncidents"
      responses:
        "200":
          description: "successful operation, only returns token if not using OAuth"
          schema:
            $ref: "#/definitions/GoodResponse"
          headers:
            Authentication:
              description: "token"
              type: "string"
        401:
          description: bad authorization token
          schema:
            $ref: "#/definitions/CreateIncidentInvalidIncidentResponse"
        "404":
          description: "userId for this incident is not found"
          schema:
            $ref: "#/definitions/CreateIncidentUserIdNotFoundResponse"
    get:
      tags:
      - "incident"
      summary: "Get incidents"
      description: "Use this to get all incidents created by a user. incidents are listed in array in order from newest to oldest created"
      operationId: "getIncidents"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: Authorization
        description: "authorization header contains bearer token"
        type: string
        required: true
      #add parameter for user id so the backend can get all incidents for a user id CANNOT BE A BODY
      - in: header
        name: User
        description: "contains user id"
        type: string
        required: true
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/GetIncidentsGoodResponse"
            
        401:
          description: "bad authorization token"
          schema:
            $ref: "#/definitions/BadResponse"
        404:
          description: "userid not found"
          schema:
            $ref:  "#/definitions/BadResponse"
            
    delete:
      tags:
      - "incident"
      summary: "Delete incident"
      description: "Use this to delete an incident by incidentId"
      operationId: "deleteIncidents"
      produces:
      - "application/json"
      consumes:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "authorization header contains bearer token"
        type: string
        required: true
      - in: header
        name: User
        description: "contains user id"
        type: string
        required: true
      responses:
        200:
          description: successful operation
          schema:
            $ref: "#/definitions/GoodResponse"
        401:
          description: bad authorization token
          schema:
            $ref: "#/definitions/BadResponse"
        404:
          description: incident not found
          schema:
            $ref: "#/definitions/BadResponse"
...
definitions:
...
  BadResponse:
    type: object
    properties:
      message:
        type: string
  GoodResponse:
    type: object
    properties:
      message:
        type: string
...
  CreateIncidents:
    type: object
    properties:
      userId:
        type: string
      incidents:
        type: array
        items:
          $ref: "#/definitions/Incident" 

    
      
        
# old incident
  Incident:
    type: "object"
    properties:
      id: 
        type: string
      location: # 1 /
        type: string
      locationOfObjects: # 2 
        type: string
      longTermPrognosis: # 3 
        type: "string"
      symptomsPresent: # 4
        type: string
      anteriorPhoto: # 6
        type: "string"
      lateralPhoto:
        type: string
      posteriorPhoto:
        type: string
      incidentYear: # 7 /
        type: "string"
      objectConsistency: # 8
        type: "string"
      gender:  # 9 /
        type: "string"
      patientAge: # 10 /
        type: "string"
      largestLength: # 11
        type: "string"
      removalStrategy: # 12
        type: string
      settingOfRemoval: # 13
        type: string
      lengthOfHospitalStay: # 4
        type: string
      lifeThreatening: # 15
        type: string
      incidentDescription: # 16 /
        type: string
      timeUntilRemoval: # 17
        type: string
      easeOfRemoval: # 18
        type: string
      objectMaterial: # 19
        type: string
      objectBasicShape: # 20
        type: string
      anesthesia: # 21
        type: string
      symptomSeverity: # 22
        type: string
      xrayOpacity: #  23
        type: string
      aceticAcid: # 24
        type: string
      other: # 25
        type: string
      dimensionality: # 26
        type: string
      additionalImagingAndSurgery: # 27
        type: string
      numberOfPieces: # 28
        type: string
      objectsIntact: # 29
        type: string
      objectCharacteristics: # 30
        type: string
      batteryLocation: # 31
        type: string
      magneticPoleDirection: # 32
        type: string
      complications: # 33
        type: string
      largestDepth: # 34
        type: string
      sucralfate: # 35
        type: string
      batteryImprintCode: # 36
        type: string
      otherShape: # 37
        type: string
      largestWidth: # 38
        type: string
      magnetType: # 39
        type: string
      numberOfObjects: # 40 /
        type: string 
      customMagnetType: # 41
        type: string
      honey: # 42
        type: string
...
  CreateIncidentInvalidIncidentResponse:
    type: "object"
    required:
      - ID
    properties:
      ID:
        type: "string"
      Long-term_prognosis:
        type: "string"
      What_material_is_the_object_made_of:
        type: "string"
      Anterior: 
        type: "string"
      Date_of_Incident:
        type: "string"
      Object_Consistency:
        type: "string"
      Gender: 
        type: "string"
      Approximate_Patient_Age:
        type: "string"
      Location_of_object:
        type: "string"
      Incident_Description:
        type: "string"
      Largest_Length:
        type: "string"
      Object_Basic_Shape:
        type: "string"
      The_object_is:
        type: "string"
      Created:
          type: boolean
          default: false
  CreateIncidentUserIdNotFoundResponse:
    type: "object"
    required:
      - ID
    properties:
      ID:
        type: "string"
      Long-term_prognosis:
        type: "string"
      What_material_is_the_object_made_of:
        type: "string"
      Anterior: 
        type: "string"
      Date_of_Incident:
        type: "string"
      Object_Consistency:
        type: "string"
      Gender: 
        type: "string"
      Approximate_Patient_Age:
        type: "string"
      Location_of_object:
        type: "string"
      Incident_Description:
        type: "string"
      Largest_Length:
        type: "string"
      Object_Basic_Shape:
        type: "string"
      The_object_is:
        type: "string" 
      Created:
        type: boolean
        default: false
...
  GetIncidentsGoodResponse:
    type: "object"
    required:
      - userId
    properties:
      userId:
        type: "string"
      incidents:
        type: array
        items:
          $ref: "#/definitions/Incident"
...
  CreateUser:
    type: object
    required:
      - id
      - name
      - email
      - degree
      - specialty
    properties:
      id:
        type: string
      name:
        type: string
      email: 
        type: string
      degree:
        type: string
      specialty:
        type: string
...
  GetUserGoodResponse:
    type: "object"
    required: 
      - userId
      - verified
      - name
      - degree
      - specialty
      - email
    properties:
      userId:
        type: "string"
      name:
        type: "string"
      degree:
        type: "string"
      specialty:
        type: "string"
      email:
        type: "string"
      verified:
        type: "boolean"
        example: false
...